Initial commit
First version of a tool that creates PDF documents from a definition in
a JavaScript object. Inspired by [pdfmake]. The implementation is based
on [pdf-lib], which is written in TypeScript and appears more robust.

Content is defined as a list of paragraphs that can include text. Text
can be provided as a plain string, as an object that includes text and
text attributes (`fontSize`, `fontFamily`, `italic`, and `bold`), or as
an array of texts. This allows for styling certain spans of a paragraph
in a different font or font size.

Fonts must be provided in the document definition. Content can select
fonts using text attributes `fontFamily`, `italic`, and `bold`.

An example is provided in the `examples/` folder, it can be generated
during development using `npm start`.

[pdfmake](https://github.com/bpampuch/pdfmake)
[pdf-lib](https://github.com/Hopding/pdf-lib)