✨ Support page margins
Support a page margin in the document definition.
To allow specifying the margin using common units, support strings that
include a number followed by a unit, as usual in CSS.
Support the units `pt` and `in` (inch), as well as the metric units `mm`
and `cm`.
These units can later also be used for other definitions such as
paddings in paragraphs.

Create functions to parse lengths with units and translate them to pt.
1 pt, defined as 1/72 inch is the default unit in a PDF document, all
layout and rendering is based on this unit.

Create a function to subtract edges from a box and ensure that the
resulting box cannot have negative widths or height.