✨ Support lineHeight attribute
Running text is usually displayed with some space between the lines.
Hence, the line height is slightly larger than the font size.

CSS has a `line-height` attribute that allows to control the height of
text lines [[1]]. The default depends on the user agent, but is roughly
a factor of `1.2`, multiplied with the font size.

Introduce a text attribute `lineHeight` that supports a multiple of the
font size and defaults to `1.2`.

As in CSS, the line height can also be set on an inline element. In this
case, the height of a row is the maximum of the line heights of all
included text spans. Set this line height as the height of the "row"
frame.

Since the text baseline is currently set to the lower edge of the row,
the additional offset is added on top of each text row. This issue will
be addressed in a subsequent change, when the text baseline is adjusted
according to the font ascend and descend.

[1]: https://developer.mozilla.org/en-US/docs/Web/CSS/line-height