✨ Support multi-line text
Each element in the document's `content` array is a paragraph.
When the text of a paragraph does not fit into a single line, it should
be broken into multiple lines.

This change implements a line-breaking algorithm that first splits text
into chunks of subsequent whitespace and non-whitespace characters and
measures these chunks separately. When trying to fit the chunks into a
given box, chunks of whitespace characters represent an opportunity for
a line break. When a chunks exceeds the max width, a line break is made
before that chunk, or, if that is not possible, after the exceeding
chunk. This ensures that text is rendered even when no line break is
possible. In this case, text will exceeds the margin.

A "paragraph" frame is introduced, that includes all rows of a
paragraph.

Co-authored-by: XintianShi <xintian@innoopract.com>