✨ Support padding and margin on paragraphs
This change adds support for the attributes `padding` and `margin` on
paragraphs.

As in the [CSS box model], a padding adds space between the edges of a
paragraph and its contents, while a margin surrounds a paragraph and
provides space between paragraphs.

Adjacent margins collapse into one, i.e. the bottom margin of one
paragraph is merged with the top margin of the next paragraph into a
single margin whose size is the maximum of the two margins.
This is also a feature of the [CSS box model].

Instead of shrinking the width of paragraphs to their content width,
set it to its available width. This allows for testing the effect of
margins, but it's also a precondition for right and center alignment.

[CSS box model]: https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Box_Model