✨ Support text color
Support `color` as another text attribute.

PDF supports colors for text and graphics in different color spaces
(grayscale,  RGB, CMYK, and other special types). Add support for RGB
colors as a start, as these colors are most common and easy to use.

Accept colors in the hexadecimal notation that is used in HTML
(`#xxxxxx`). Also support a small set of named colors, mainly as
shorthands for experiments. This set can later be customized.

By introducing a color type, this change also lays the foundation for
drawing graphical shapes with colors.