✨ Support links to URLs in text
This change adds support for `link` as an additional text attribute.
An annotation of type `Link` is created in the page for each link.

A single link can span multiple text segments, e.g. when the font
changes within the text range of a link. In this case, the
implementation ensures that only a single link annotation is created.

See Adobe PDF 1.7, section 8.4.5, Annotation Types
https://archive.org/details/pdf1.7/page/n621/mode/2up