✨ Support drawing rectangles, lines, and polylines
To allow drawing basic graphics shapes into paragraphs, e.g. for markers
or annotations, this change introduces a new attribute `graphics` in
paragraph objects. This attribute accepts an array of graphics objects.
Supported types of graphics objects are `rect`, `line`, and `polyline`.

The coordinates of graphics objects are interpreted relative to the
padding of the paragraph. The origin of the coordinate system is the
top left corner, similar to [SVG] and the [Canvas API].

Parse the graphics objects before processing to ensure that no invalid
attributes can lead to unexpected errors during layout and render.
The parse function picks only the relevant members of a given object and
check their types.

Graphics objects are included in the paragraph frame's `objects`. Render
the drawable objects of a frame before child frames to draw text on top
of drawings.

[SVG]: https://developer.mozilla.org/en-US/docs/Web/SVG/Tutorial/Positions
[Canvas API]: https://developer.mozilla.org/en-US/docs/Web/API/Canvas_API