✨ Support text alignment
Previously all texts are aligned to the left by default. This change
adds support for the attribute textAlign on paragraphs. As in the
[CSS text-align], while setting textAlign to center, the texts
including the paddings are aligned to the center. While setting
textAlign to right, the texts including paddings are aligned to the
right. While `textAlign` is not set, align texts to the left.

[CSS text-align]: https://developer.mozilla.org/en-US/docs/Web/CSS/text-align