✨ Introduce defaultStyle attribute
Add support for an attribute `defaultStyle` in the document definition.
This attribute accepts an object of text attributes that are applied as
defaults to all texts in the document.

The attribute is called "defaultStyles" instead of "defaultTextAttrs"
as it may be used for other kinds of style attributes later, e.g. those
that affect graphics. The term "style" is also common in CSS.

This change also improves type checks for content:
- Errors in paragraphs are preceded with the paragraph number, e.g.
  `Invalid value for "paragraph #n"`.
- Top-level text attributes in paragraphs are now type checked.
- Attributes `fontSize` and `lineHeight` must be non-negative if
  present.