✨ Support multi-page documents
So far, pdf-maker could only create a single-page document.
To support longer documents, distribute content across multiple pages.

Let `layoutPage` stop when a paragraph does not fit on the page, and
return the page frame and the remaining paragraphs. Introduce a function
`layoutPages` that creates new pages for the remaining paragraphs.

This change only inserts page breaks between entire paragraphs, not
between rows. This is going to be improved later on.

Before, the page size used to be read from the page. Since the pages are
now created during rendering, this is not possible anymore. Set the page
size to A4 explicitly for now.