✨ Support columns
To allow arranging paragraphs horizontally, this change introduces a
content type `Columns` as an alternative to `Paragraph`.
With this change, content is no longer a sequence of paragraphs, but a
sequence of blocks, that can be either of type `Columns` or `Paragraph`.
Other type of blocks may be added later to support different layouts.

The attributes `width`, `height`, and `margin` are extracted from
`Paragraph` and declared as block attributes. They should be supported
for all kinds of blocks, regardless of their type.

A `Columns` element is an object with a `columns` array. Columns are
again blocks. The columns layout respects the `width` attribute of
individual columns and distributes the remaining space evenly across the
columns that don't have a fixed width.

Since columns are not free floating, but arranged in a grid with
pre-calculated widths, margins of adjacent columns are not collapsed.
This would also not be very intuitive. A "column gap" attribute would
solve the same problem but in a clearer way. This attribute is not yet
implemented but may be added later.

The `parseBlock` and `parseColums` function are implemented next to
`parseParagraph` in `text.js`. This is an unfortunate place, but code
should be moved in a separate change.