✨ Support headers and footers
This change adds support for static and dynamic page headers and
footers. The optional attributes `header` and `footer` accept either a
content block or a function that returns a content block for each page.
If a function is given, it is called with a page info object including
the page number, total page count, and the page size.

In order to pass the total page count to functions, these functions must
be called after the layout of the main content is complete. Parsing
dynamic content during the layout process requires moving the default
page size and margin to `layout.ts`. Calculating margin and boxes should
be a responsibility of the layout module anyway.

To include frames for header and footer in the objects returned by
`renderPage`, return a sequence of page objects instead of frames.
This makes the `createPage` function obsolete. The `renderPage` function
now creates the PdfPage and adds it to the page object.