🐛 Rename util function "pick" to "getFrom"
This function retrieves a single value from an object, checks its type
and throw an error that includes the attribute name.
The function name "pick" is not a good choice as this term is associated
with picking a subset of the attributes of an object (functions of this
name exist in lodash and underscore).

This change renames the function to `getFrom` which makes the purpose
clearer.