✨ Support JPEG images
Images need to be embedded as objects in the PDF document. In order to
allow reusing the same image at different locations within the document
without multiplying their footprint, require images to be registered
before use.

When an explicit `width` or `height` is specified on the block element,
fit the image into the given bounds. This appears to be the most useful
strategy. Other resize strategies such as "fill" or "stretch" can be
implemented later when there is a demand.

Include a public domain image in the sample. The original image is
taken from [publicdomainpictures].

[publicdomainpictures]: https://www.publicdomainpictures.net