♻️ Improve error messages for type checks
When type checking input data, a util function is used to create a
`TypeError` with a message following the common pattern
"Expected {this}, got: {that}".
Rename this function to `typeError()`, export it, and replace manually
created type errors with this function where possible to align all error
messages to the same pattern.

Including the actual value in the error message concealed certain
values. For example, an empty array would be represented by the empty
string, objects as `[object Object]`. Numbers couldn't be distinguished
from numeric strings. This can lead to confusing error messages such as
"Expected number, got 23".

To avoid this confusion, implement custom print functions for common
types such as strings, arrays, and objects.

Support using `optional()` without arguments to match the behavior of
`required()` without arguments (return the input value unmodified).
Remove a useless optional chaining operator in the implementation of
`getFrom()`.