✨ Support internal links
Support anchors on block elements by adding a unique `id` attribute.
Anchors are included as [named destinations] in the `Dests` element of
the document's [name dictionary].

Allow links to those named destinations by setting the text attribute
`link` to a hash sign (`#`) followed by the id of the referenced
element.
This approach seems simple and intuitive, it matches the way that links
to internal anchors are defined in HTML documents.
When a link starts with a hash sign, assume it's an internal link and
render it using a [goto action] to this named destination.

Base expectations in unit tests on the `toString` methods of PDF types
to assert the actual PDF output. This provides more confidence than
assertions that are based on mocks.

[named destinations]: See Adobe PDF 1.7, section 8.2.1, Destinations
https://archive.org/details/pdf1.7/page/n579/mode/2up

[name dictionary]: See Adobe PDF 1.7, section 3.6.3, Name Dictionary
https://archive.org/details/pdf1.7/page/n149/mode/2up

[goto action]: See Adobe PDF 1.7, section 8.5.3, Action Types
https://archive.org/details/pdf1.7/page/n651/mode/2up