🐛 Align image definition format with API type
According to the type definition in `content.ts`, images are registered
by a name and an object with a `data` attribute:

```js
{
  images: {
    foo: { data: <image data> }
  }
}
```

However, the implementation expected the image data as value directly.
Using an object allows for additional attributes, such as image type,
default size, should they become necessary.

Adjust the implementation and sample to the type definition.