💥 Separate images from text paragraphs, revise API
Separate the layout of image blocks from the layout of text paragraphs
to decouple both implementations.

With this change, a block can contain either an image or text, but not
both. The previous implementation would render text on top of the image
when both were present in a block. However, there was no way of
positioning text on the image, so this feature was rather useless.
Some future version should offer means of layering content on top of
other content instead.

When neither `width` nor `height` is specified, the image is not scaled.
However, if the image width exceeds the maximum available width, scale
it down to fit onto the page.

Extract the `layoutImage` function to a module of its own and add
some unit tests.