🐛 Fix position of anchors, rename dest to anchor
Anchors are called "named destinations" in PDF land [1]. Rename the
internal object type `dest` to `anchor` which is a bit easier to
decipher.

When a frame contains an `id`, an anchor should be created that points
to the top-right corner of this frame. However, anchors for paragraphs
were mistakenly offset by the padding. Frames for rows and columns were
even offset by the coordinates of the frame. Since the anchor is
represented by an object, whose positions are relative to the frame,
anchor objects should normally not have an offset.

Remove this offset.

[1] https://archive.org/details/pdf1.7/page/n581/mode/2up