♻️ Introduce internal type Document
Embedding fonts and images into the PDF document should be up to the
document module. Move the code into the `createDocument` function.

Return a new type `Document` that wraps pdf-lib's `PDFDocument` and adds
the lists of registered fonts and images that were previously tracked in
a `Resources` object. This new type can be extended as needed to capture
all document-related internal attributes.

Remove the type `Resources` that has become obsolete by this change.