♻️ Align type checks, introduce isObject
For historic reasons, the functions `parseColor()`, `parseLength()` and
`parseEdges()` return `undefined` for null and undefined values. This
special handling is not needed anymore since both functions are always
wrapped in an `optional()` call. Remove it.

Use the `typeError` function to creating type errors in `parseColor()`
to align with error messages for other types.

Introduce a function `isObject` and use it where appropriate. This
function handles some edge cases that haven't been handled properly
before:

* returns `false` for `null` (`typeof null -> 'object'`)
* returns `false` for arrays (`typeof [] -> 'object'`)
* returns `false` for objects with constructors other than `Object`