✨ Support different page sizes
The page size of the PDF was fixed to A4, portrait. This change allows
creating PDF documents with different page sizes and orientations.

Introduce a new attribute `pageSize` in the document definition that
allows to set the size for all pages in the document. Different sizes
for individual pages are not yet supported.
The page size is accepted either as an object with `width` and `height`
attributes or a string with a well-known paper format, e.g. "A4". Format
strings are case sensitive, i.e. `Letter` is supported, but `letteR` is
not. Page sizes have been copied from pdf-lib to ensure the API remains
constant.

Also support an attribute `pageOrientation` to set the page orientation
to portrait or landscape. When this attribute is set, the given page
size is flipped if necessary to match the given orientation.