✨💥 Introduce imageAlign, center images by default
Images are scaled proportionally to fit in their given frame. When an
image does not fill its block, it was always aligned to the left.
While left alignment is the natural default for text, it does not seem
like a good default for images.

Center-align images by default and add a property `imageAlign` to set
the alignment.