♻️ Extract generation of anchors from layout functions
When a block contains the attribute `id`, an anchor object is created in
the frame. The anchor always points to the top left corner of the frame,
regardless of the type of content.

Remove the redundant implementation from individual layout functions and
create the anchor object in a central place (`layoutBlock`).