♻️ Introduce type check function for dynamic types
Dynamic types accept either a fixed value or a function that returns a
value. If a function is provided, is is evaluated multiple times, e.g.
once for every page, with different arguments.

Introduce a function `dynamic()` to wrap a type definition and return a
dynamic type definition. When applied, this type definition returns a
resolve function that returns the resolved value.

When applied to a fixed value, the type check is performed immediately
and the error message includes the attribute name. When applied to a
function, the returned value is checked after the resolve function is
called. At that point, the attribute name is not available anymore. To
name the source in the error message, an additional parameter `name` can
be set.

Use this dynamic type for headers and footers.