✨💥 Support graphics functions on all block types
The `graphics` attribute was only supported on text blocks. The drawing
area was inside the padding of the block, which was unfortunate when
drawing border elements around a frame.

Support the `graphics` attribute on all blocks and let the drawing area
cover the entire frame.

Support functions to be supplied to the `graphics` attribute. Functions
are called with an info object that contains the final size of the
rendered block. This allows for drawing exactly around the edges of a
frame.

Remove code to shift graphic elements which was only needed to apply the
padding.

Rename the `graphics` module to `read-graphics` as it now solely reads
graphics objects.