♻️ Render graphics using low-level operations
We used pdf-lib's higher-level graphics functions `drawLine`,
`drawRectangle`, and `drawSvgPath` to render graphics objects.
These functions have some limitations. For example, they do support
`lineCap`, but not `lineJoin`. They also do not support "Automatic
Stroke Adjustment", a PDF feature that aligns fine lines with the raster
of the rendering device to avoid blur.

Implementing the rendering of graphics objects using low-level
operations provides more flexibility and allows for a more efficient
rendering of bulk operations. It also avoids additional translations to
pdf-lib's attribute names.

Implement drawing operations in a new module, `render-graphics`.
For rectangles, use the shorthand operator `re` [[1]].

Stroke paths of objects that have a `strokeColor`, fill paths of objects
that have a `fillColor`. As a sensible default for objects that have
neither, stroke the path if it's a line, and fill it otherwise. PDF will
use black as the default color.

[1]: See Adobe PDF 1.7, section 4.4.1, Path Construction Operators, https://archive.org/details/pdf1.7/page/n225/mode/2up