✨ Support lineCap and lineJoin attributes
Support the attributes `lineCap` and `lineJoin` on graphics objects to
allow for smooth edges and joins.

In SVG, the corresponding attributes are called `stroke-linecap` and
`stroke-linejoin`. In HTML Canvas, it's `lineCap` and `lineJoin`. Also
the PDF spec only talks about "line cap" and "line join". Obviously,
these attributes affect only the strokes, not the filled regions, to the
`stroke` prefix seems redundant.

The attribute names `butt`, `round`, and `square` for `lineCap`, as well
as `miter`, `round`, and `bevel` for `lineJoin` are established in both
SVG and HTML Canvas, and they match the terms used in the PDF spec.