♻️ Render text using low-level operations
Text was rendered using pdf-lib's higher-level function `drawText`. This
function also measures texts and wraps lines, things that are already
done by the our code.

Rendering text using low-level operations gives us finer control and
avoids unnecessary operations. By rendering all spans of a frame in one
function, we can keep track of the graphics state and don't need to set
the same attributes again and again.

pdf-lib created a new entry in the page's fonts dictionary each time a
font was used. This is not necessary. To avoid multiple entries for the
same font, introduce a function `getFont` that keeps track of the fonts
registered with the page and only registers a font once.