✨ Support opacity for graphics
Support additional attributes `lineOpacity` and `fillOpacity` for
graphics shapes.

Opacity cannot be changed by an operator in the content stream, it
requires an entry in a *graphics state parameter dictionary* [1].
Introduce a function `getPageGraphicsState` to set a graphics state
dictionary on the page when it is first used. Reuse this graphics state
for subsequent operations with the same opacity parameters.

For consistency, also rename functions `getFont` and `getImage` to
`getPageFont` and `getPageImage`. All these functions get and cache an
object in the scope of a page.

[1]: See Adobe PDF 1.7, section 4.3.4 Graphics State Parameter Dictionaries
https://archive.org/details/pdf1.7/page/n217/mode/2up