♻️ Extract read functions for blocks
In the beginning, there were only paragraphs. Functions to read other
kinds of blocks were added to the same module. So the `text` module
became a collection of functions for reading blocks and operations on
texts.

Move the read functions to a new module, `read-block` and rename the
functions to start with `read` instead of `parse`, following an
established naming pattern.