🐛 Vertically center links in text rows
Text is vertically centered in text rows. However, links were still
positioned at the upper edge of the row. With the standard line height
of 1.2, this offset was hardly noticeable, but for larger values it is.

Take the vertical offset into account when creating link objects.
Slightly adjust the offset calculation to avoid rounding errors which
would break tests.

In most cases, positions in PDF Maker are relative to the top left
corner of a frame. Follow this convention also for link objects.