⚡️ Aggregate text rows in a block
For each line of text in a block, a separate frame was created. This
lead to initializing a new PDF text object for each row, including
setting font and text color again for each row.

Instead of creating child frames for the rows of a text block, combine
all lines of a block in a single text object and render them as a single
text object. With this change, the text matrix is only set once for each
block, also font and color are only set once at the start of a text
block and when they change.

Include width, height, and baseline in text row objects and draw guides
based on these objects.