💥 Introduce EmptyBlock, require text and image attrs
The attribute `text` in `TextBlock` was optional to allow for empty
blocks, e.g. those that act as spacers or contain only graphics.
This pattern had been copied to `ImageBlock` whose attribute `image` was
also optional.

Make support for empty blocks explicit by introducing a new block type
`EmptyBlock`. In turn, make the attributes `text` in `TextBlock` and
`image` in `ImageBlock` required. There is no point in a text block
without text or an image block without an image.