♻️ Drop type attribute from render frames
The only purpose of the `type` attribute on render frames was to select
the color for guides. With the introduction of rows guides, which are
not based on frames, the idea of rendering guides based only on frames
has proved to be inappropriate.
With the exception of the red page guides, the different colors for
frames (blue and gray) has not proven helpful.

Introduce a dedicated function `createPageGuides` that draws the red
guides around the page's print space and use the existing function
`createFrameGuides` for all other frames. Draw the guides for all
regular frames in gray, reserve colors for other types of guides (e.g.
paddings / margins).

Remove the `type` attribute from `Frame`, so that all frames are now
equal.