♻️ Extract handling of padding and fixed block size
Layout calculations for different block types had to take the padding
of the block as well as fixed width and height into account. The logic
for both is the same for any kind of block.

Handle the `padding`, `width`, and `height` attributes of a block in
`layoutBlock`. With this change, layout function for the individual
types of block content only need to layout the content in a given box.
This box already includes padding and fixed size.

Only `layoutImageContent` still checks the presence of a `width` and
`height` attribute on the block to select the scaling method.