✨ Display margins and paddings with guides
To help understanding the effects of margins and paddings, make them
visible when showing guides.

Draw a semi-transparent background on areas that are reserved by block
paddings and margins. Use a subtle yellow color for margins and a
blueish one for paddings, as usual in browser dev tools. The
transparency also helps to distinguish collapsed margins.

Represent the page margin using a thick line outside the page's print
space. The thin red line that was shown before was always shadowed by
the edges of block guides and, since margins are outside, the thick line
might even be more intuitive.