✨ Support block attributes to control page breaks
Sometimes, auto-inserted page breaks are unfortunate, e.g. when the
headline of a new section appears at the very bottom of a page,
separated from the content it relates to (an [orphan]).

To give content definitions more control over page break locations,
support the attributes `breakBefore` and `breakAfter` on top-level
blocks. They can be used to control whether a page break should or
should not be inserted before or after a block. These attributes are
modeled after the CSS attributes [break-before] and [break-after] which
replace the deprecated `page-break-before` and `page-break-after`.
The new attributes control not only page breaks, but also column breaks.
Even though the latter is not yet required, we may want to extend these
attributes later.

Currently, only the attributes `auto`, `avoid`, and `always` are
supported, with `auto` being the default.

In case of a conflict, e.g. when a block with `breakAfter: 'avoid'` is
followed by a block with `breakBefore: 'always'`, let `always` take
precedence and insert a page break.

When guides are enabled, indicate enforced page breaks by a small
horizontal bar above the frame on the left (for `breakBefore: 'always'`)
or below the frame on the right (for `breakAfter: 'always'`).
Also indicate `avoid` by a small dot on the upper or lower edge of the
frame.

[orphan]: https://en.wikipedia.org/wiki/Widows_and_orphans
[break-before]: https://developer.mozilla.org/en-US/docs/Web/CSS/break-before
[break-after]: https://developer.mozilla.org/en-US/docs/Web/CSS/break-after