🩹 Add a trailing newline to the PDF
We've seen problems with a tool that appends a revision to the PDF.
The PDF data created by pdf-lib ends with `%%EOF`, without a newline, so
appending lines directly created a syntax error.

It's seems to be good practice to terminate textual files with a newline
character, so add one.