✨ Add a file identifier to the PDF
PDF documents may include an optional "file identifier" [1]. Although
this ID is optional, it is required for digitally signed PDF/A
documents.

Generate an ID as a SHA-256 hash over the current timestamp and the
document metadata. The DOM API `TextEncoder` and `crypto` is also
available in recent versions of Node.js and Deno.

At some point, we may also want to include content in the hash, which
may be a bit more tricky. However, the current hash should be
sufficiently unique already.

[1]: See Adobe PDF 1.7, section 10.3 File Identifiers
https://archive.org/details/pdf1.7/page/n845/mode/2up