✨ Support custom data in document catalog
PDF applications may include custom attributes in any object in the PDF
document, except for the file trailer dictionary. Custom keys should be
prefixed with `XX` [1].

The `info` attribute already allows for custom meta data to be added to
the `Info` dictionary, however, only strings are supported in this data
structure.

To allow for including larger blocks of binary data in a document,
provide another attribute `customData` that accepts an object whose
values can be strings or Uint8Arrays. Data added to this attribute will
be included as stream objects and referenced in the *document catalog*.

Do not enforce the `XX` prefix, but leave the name to the application
for maximum flexibility.

[1] See PDF 1.7, Appendix E - PDF Name Registry
https://archive.org/details/pdf1.7/page/n1017/mode/2up