✨ Support text rise for superscript and subscript
This change adds support for a new optional text attribute named `rise`.
This attribute specifies a vertical offset that shifts the text baseline
up or down from its default location. Adjustments to the baseline are
useful for drawing superscripts or subscripts.

The text rise is rendered using the PDF operator `Ts` [1].

[1]: See Adobe PDF 1.7, section 5.2.6 Text Rise
https://archive.org/details/pdf1.7/page/n402/mode/1up