💥 Make page margin relative to header/footer, bump to 0.4.0
The height of the header and the footer can be dynamic, but the page
content used to be positioned relative to the edges of the page which
could lead to content overlapping the header or footer.

This commit makes the page margin relative to the header and footer to
let the page contents adjust to a dynamic header and footer height. The
larger the header and footer, the less space is available for the page
content. It is still possible to define a static layout by setting a
fixed height on the header and footer.

This change requires to layout the header and footer first. However, at
this point, the total page count is not yet known. Therefore, the header
and footer are rendered twice, once to determine their size, and once
again after all pages have been laid out to insert the total page count.
This is only done if the header or footer is defined as a function.

Bump version to 0.4.0 to account for the breaking change.