✨ Support dashed and dotted lines
This commit adds support for a new property `lineDash` on all graphics
shapes. This property accepts the dash pattern as an array of numbers,
each of which defines the length (in pt) of a dash or gap, starting with
the first dash. Dotted lines can be drawn by setting `lineCap` to
`round` and using `0` as the length of the dash.

The same method of specifying dash patterns is actually used in SVG [1]
and the HTML Canvas API [2]. PDF also supports a dash *phase*, which
is not exposed by this commit [3].

[1]: https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/stroke-dasharray
[2]: https://developer.mozilla.org/en-US/docs/Web/API/CanvasRenderingContext2D/setLineDash
[3]: See Adobe PDF 1.7, section 4.3.2, Line Dash Pattern
https://archive.org/details/pdf1.7/page/n217/mode/1up