🐛 Fix text rise not always reset
When a text rise was set in the last chunk of a text content object, it
was not reset at the beginning of the next text object. This caused the
text rise to be applied to the rest of the page.

Text rise is a text state operator [1] that is retained across text
objects in a single content stream, i.e. a page.

Consequently, this commit maintains the text state across text objects
within a page.

[1]: See 5.2 Text State Parameters and Operators
https://archive.org/details/pdf1.7/page/n395/mode/1up