✨ Support drawing SVG paths
This commit adds full support for SVG paths in `graphics`. To draw an
SVG path, an element of type `path` accept an SVG path in a parameter
`d`, similar to the SVG `<path>` element. Paths are expected in the
syntax used by the `d` attribute of the SVG `<path>`. All path commands
as described in [1] are supported.

[1] https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/d#path_commands