♻️ Introduce util function compact
We sometimes include conditional values in an array and then filter out
the falsy values using `filter(Boolean)`. This approach requires the use
of a type assertion (using `as`) to get the correct type.

This commit introduces a util function `compact`, inspired by the lodash
function of the same name, that reflects the filter in its return type.