♻️ Dedicated result type for content layout functions
Layout function for block content do not need to return the position of the

Functions that layout the *content* of a block do not need to return its
position. The position is fixed to the top left corner of the padding by
now. Also, the width is currently fixed to the available width.

This commit replaces the `Partial<Frame>` return type with a dedicated
type that requires only the `height` attribute.