♻️ Correct return type for pickDefined function
The `pickDefined` function strips all `undefined` values from a given
object of type `T`. The return type of the function was `Partial<T>`.
However, since only optional fields can be `undefined`, this function
does not change the type.

This commit adjusts the return type to reflect the input type. This
helps to remove some type assertions.