✨💥 Support page breaks inside text blocks
This commit adds support for page breaks inside text blocks. A new
`breakInside` property is added to the `TextBlock` type, similar to the
`break-inside` CSS property. This property can be used to control
whether a row block can be broken across pages or not. The default value
is `auto`. Setting it to `avoid` will prevent page breaks inside the
block.