✨ Support inserting an extra block after page break
When a page break occurs inside a block with `rows`, it may be desirable
to insert an extra block just before the first element on the new page,
e.g. a table header.

This commit introduces a new `insertAfterBreak` property on `RowsBlock`
that allows specifying a block to insert after a page break.