✨ Support transformations on graphics
Transformations allow to move, scale, or rotate elements without having
to recalculate their coordinates. This is particularly useful when
reusing complex graphical elements such as SVG paths.

This commit introduces the new attributes `translate`, `scale`,
`rotate`, `skew`, and `matrix` to all `Shape` types. These attributes
support the same transformation functions as the [transform] attribute
in SVG.

In SVG, various transformation functions must be encoded in a single
string, e.g. `transform="translate(10 20) rotate(-10 50 25)"`. It seems
inconvenient to assemble these strings in a program and also to parse.
For this reason, separate attributes are used instead of `transform`.

The numbers in the transformation matrices are rounded to 6 decimal
places. This seems to provide a good balance between precision and file
size. Event for very large values, this precision should be sufficient.
For example, for a base value of 10,000 pt which is about 3.5 meter, an
inaccuracy of 0.000001 would result in a difference of 0.01pt.

[transform]: https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/transform