✨ Support dynamic page margins
Pages with odd numbers ("recto") often need a larger margin on the left
than on the right. For duplex printing, the margins need to be reversed
on the even pages ("verso").

This commit adds support for dynamic page margins by accepting a
function in the `margin` attribute of a document definition.