♻️ Decouple layout from PDF lib types
The layout phase relied on the `pdf-lib` types `PDFFont` and `PDFImage`.
These types can only be created by embedding a font or an image into the
PDF. In preparation of loading font and image resources on demand, this
commit decouples the layout process from these `pdf-lib` types and uses
only `Font` and `Image` during layout.

Before layout, fonts are now only *loaded*, i.e. the `fontkit` font is
being created. To calculate text sizes, the layout now relies only on
`fontkit` instead of a `PDFFont` instance. The fonts are embedded into
the PDF only in the render phase.

Likewise, images are loaded before layout and embedded into the PDF in
the render phase. They need to be loaded in advance to get their size.