🩹 Do not set default Producer and Creator
Previously, the "Producer" and "Creator" fields in the PDF Info
dictionary have been set by pdf-lib. They could be overridden in the
document config, but there was no way to remove these fields. Both
fields are optional, so there is no need to set them by default.

This commit removes the default values. Users can set them as needed in
the document definition.